+++ src/index.js
@@ -1,50 +1,10 @@
-//Ajustar texto
-//juan camilo gonzalez
-console.log('hola 1')
-
-const name = 'Fredy';
-const lastname = 'Ballesteros';
-
-let inputdata = 8;
-inputdata = 4;
-
-console.log(name, lastname, inputdata)
-
-if (true){
-    let inputdata = 78
-    console.log(inputdata)
-}
-console.log(inputdata)
-
-const fullname = `${name} ${lastname}`
-console.log(fullname)
-// Fredy Ballesteros
-//Desarrollo SW2
-
-let i = 0
-i++
-//cesar solano
+import React from 'react'
+import ReactDOM from 'react-dom'
+import './index.css'
+import Primercompon from './Primercompon'
 
 
-const person ={
-    name: 'Fredy',
-    lastname: 'Ballesteros',
-    age:43
-}
 
-console.table(person);
-const person1 ={
-    name: 'Fredy',
-     lastname: 'Ballesteros',
-     age:43,
-     address: {
-        city: 'Cali',
-        zip: 6544,
-        lat: 14.24,
-        ing: 34.52
-     }
-}
+const divRoot = document.querySelector('#root')
 
-console.log(person1)
-const person3 = { ...person1}
-console.log(person3)
+ReactDOM.render(<Primercompon  priProp="123" propText="Soy texto" /> ,divRoot)
